+++ NEWS.md
@@ -3,6 +3,8 @@
 
 * `install_github()` now uses credentials from the git credential store,
   if `GITHUB_PAT` and `GITHUB_TOKEN` are not set.
+* `available_packages()` now supports the use of ellipsis, allowing additional arguments such as `headers` to be 
+  passed to `available.packages()` as well as `install.packages()` (@AshesITR, #682).
 
 # remotes 2.4.2